backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.6.1
websockets==14.1
redis==5.2.0
alembic==1.14.0
email-validator==2.2.0
googlemaps==4.10.0
stripe==11.1.1
